refactor: move install command to sharpie/install and fix review issues
